remove jsDecoupledAppDelegate code, replace w/ categories, remove unnecessary headers
