--- ios/Artsy/App/ARAppActivityContinuationDelegate.h
@@ -1,4 +1,0 @@
-#import <JSDecoupledAppDelegate/JSDecoupledAppDelegate.h>
-
-@interface ARAppActivityContinuationDelegate : NSObject <JSApplicationActivityContinuationDelegate>
-@end
