--- ios/Artsy/App/ARApplicationShortcutItemDelegate.h
@@ -1,5 +1,0 @@
-#import <JSDecoupledAppDelegate/JSDecoupledAppDelegate.h>
-
-@interface ARApplicationShortcutItemDelegate : NSObject <JSApplicationShortcutItemDelegate>
-
-@end
