+++ ios/Artsy/App/AppDelegateCategories/AppDelegate+ActivityContinuation.m
@@ -1,20 +1,8 @@
-#import "ARAppActivityContinuationDelegate.h"
-
-#import "ARUserManager.h"
-#import "ArtsyAPI.h"
-
+#import "AppDelegate+ActivityContinuation.h"
 #import <CoreSpotlight/CoreSpotlight.h>
-#import "AREmission.h"
 #import <React/RCTLinkingManager.h>
 
-static  NSString *SailthruLinkDomain = @"link.artsy.net";
-
-@implementation ARAppActivityContinuationDelegate
-
-+ (void)load
-{
-    [JSDecoupledAppDelegate sharedAppDelegate].activityContinuationDelegate = [[self alloc] init];
-}
+@implementation ARAppDelegate (ActivityContinuation)
 
 - (BOOL)application:(UIApplication *)application willContinueUserActivityWithType:(NSString *)userActivityType;
 {
