+++ ios/Artsy/App/AppDelegateCategories/AppDelegate+Echo.h
