+++ ios/Artsy/App/AppDelegateCategories/AppDelegate+Emission.h
