+++ ios/Artsy/App/AppDelegateCategories/AppDelegate+Emission.m
@@ -1,34 +1,9 @@
-#import "ARAppDelegate+Emission.h"
-
-#import "ARUserManager.h"
-#import "Artist.h"
-#import "ArtsyEcho.h"
-#import "Gene.h"
-#import "ArtsyAPI+Following.h"
+#import "AppDelegate+Emission.h"
 #import "ArtsyAPI+Notifications.h"
-#import "ARDispatchManager.h"
-#import "ARNetworkErrorManager.h"
-#import "ARAppConstants.h"
-#import "AROptions.h"
-#import "ARMenuAwareViewController.h"
-#import "ARAppNotificationsDelegate.h"
-#import "ARAugmentedVIRSetupViewController.h"
-#import "ARAugmentedRealityConfig.h"
-#import "ARAugmentedFloorBasedVIRViewController.h"
-#import "ARDefaults.h"
-#import "ARAppStatus.h"
 #import "ARRouter.h"
-#import "AROptions.h"
-
-#import "Keys.h"
 #import "AREmission.h"
 #import "ARTemporaryAPIModule.h"
 #import "AREventsModule.h"
-#import <SDWebImage/SDImageCache.h>
-
-#import <React/RCTUtils.h>
-#import <React/RCTDevSettings.h>
-#import <objc/runtime.h>
 #import "ARUserManager.h"
 #import "User.h"
 #import "ARAnalyticsConstants.h"
