+++ ios/Artsy/App/AppDelegateCategories/AppDelegate+Notifications.h
@@ -1,0 +1,8 @@
+#import "AppDelegate.h"
+#import <UserNotifications/UNUserNotificationCenter.h>
+
+ @interface ARAppDelegate (Notifications) <UNUserNotificationCenterDelegate>
+
+ - (void)applicationDidReceiveRemoteNotification:(NSDictionary *)userInfo inApplicationState:(UIApplicationState)applicationState;
+
+ @end
