+++ ios/Artsy/AppDelegate.h
@@ -1,18 +1,10 @@
-#import <JSDecoupledAppDelegate/JSDecoupledAppDelegate.h>
 #import <Expo/Expo.h>
 #import <RCTAppDelegate.h>
 #import <BrazeKit/BrazeKit-Swift.h>
-#import <UserNotifications/UNUserNotificationCenter.h>
 
 @class ARWindow, ArtsyEcho;
 
-// This class, and in fact the complete JSDecoupledAppDelegate class, is not used during testing.
-// The test app delegate class is ARTestHelper and is responsible for seting up the test env.
-//
-// When testing the various decoupled app delegate classes, simply use the shared app delegate
-// (`[JSDecoupledAppDelegate sharedAppDelegate]`) to perform your tests on.
-
-@interface ARAppDelegate : RCTAppDelegate <JSApplicationStateDelegate>
+@interface ARAppDelegate : EXAppDelegateWrapper
 
 + (ARAppDelegate *)sharedInstance;
 + (Braze *)braze;
@@ -21,11 +13,21 @@
 @property (strong, nonatomic) ARWindow *window;
 @property (strong, nonatomic) UIViewController *viewController;
 
+typedef NS_ENUM(NSInteger, ARAppNotificationsRequestContext) {
+     ARAppNotificationsRequestContextLaunch,
+     ARAppNotificationsRequestContextOnboarding,
+     ARAppNotificationsRequestContextArtistFollow,
+     ARAppNotificationsRequestContextNone
+};
+
 @property (strong, nonatomic, readonly) NSString *referralURLRepresentation;
 @property (strong, nonatomic, readonly) NSString *landingURLRepresentation;
 
 /// The Artsy echo instance for feature flags, and url routing etc
 @property (nonatomic, readwrite, strong) ArtsyEcho *echo;
+
+// Notifications Delegate
+@property (nonatomic, readwrite, assign) ARAppNotificationsRequestContext requestContext;
 
 @end
