+++ ios/ArtsyTests/App_Tests/AppDelegate+ContinuationTests.m
@@ -1,18 +1,17 @@
-#import "ARAppActivityContinuationDelegate.h"
 #import "ARUserManager.h"
 #import "AppDelegate.h"
 
 #import <CoreSpotlight/CoreSpotlight.h>
 #import "AREmission.h"
 
-SpecBegin(ARAppActivityContinuationDelegate);
+SpecBegin(ARAppDelegateActivityContinuation);
 
 __block UIApplication *app = nil;
 __block id<UIApplicationDelegate> delegate = nil;
 
 beforeEach(^{
     app = [UIApplication sharedApplication];
-    delegate = [JSDecoupledAppDelegate sharedAppDelegate];
+    delegate = [[ARAppDelegate alloc] init];
 });
 
 it(@"does not accept unsupported activities", ^{
